Created Validation class and ValidationTest class. Moved validation logic from our Board class to Validation class.
