+++ test/validation_test.rb
@@ -1,0 +1,21 @@
+require 'minitest'
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/ship'
+require './lib/cell'
+require './lib/board'
+require './lib/validation'
+
+class ValidationTest < Minitest::Test
+
+  def setup
+    @submarine = Ship.new("Submarine", 2)
+    @cruiser = Ship.new("Cruiser", 3)
+    @board = Board.new
+    @validate = Validation.new()
+  end
+
+  def test_it_exists
+    assert_instance_of Validation, @validate
+  end
+end
